refactor(orgchart): extract PositionCard component
Move PositionCard rendering logic from OrgChartPage.tsx into a new PositionCard.tsx component. Removes getInitials and renderPositionCard functions from the page, reducing its size by 107 lines while improving component reusability and testability. The new PositionCard component accepts callbacks for getChildren and getLevelColor, enabling recursive rendering of organization hierarchies.
